Added Save Prompt On Exit & Fixed Toolbar Names
I added a prompt so that when the user attempts to exit the program, they are prompted to save their changes. This is a critical feature for any notepad because the humans that use these things don't always think before they close. I also fixed some of the pop up labels on the toolbar seeing as some of them were not named correctly. The background color and the font color labels were also switched so I fixed dat too.